+++ tests/programs/test_program_ahs_braket.py
@@ -15,6 +15,7 @@ Unit tests for loading, encoding, and decoding AHS programs using Amazon Braket.
 
 """
 import json
+from unittest.mock import Mock
 
 import numpy as np
 import pytest
@@ -274,6 +275,13 @@ def test_braket_ahs_wrapper_equality(ahs_program, ahs_program_local_detuning, de
 
 
 def test_braket_ahs_wrapper_bad_type():
-    """Test that creating a BraketAHS wrapper with a bad type raises a TypeError."""
+    """Test that creating a BraketAHS wrapper with a bad type raises a ProgramTypeError."""
     with pytest.raises(ProgramTypeError):
         BraketAHS(())
+
+
+def test_braket_ahs_program_transform(ahs_program):
+    """Test that the BraketAHS transform method does not modify the program."""
+    program = BraketAHS(ahs_program)
+    program.transform(device=Mock())
+    assert program.program == ahs_program